Перенос модели User в папку Models
